+++ src/application.rs
@@ -54,6 +54,7 @@ use crate::checkers::wp_plugins::elementor::ElementorChecker;
 use crate::checkers::wp_plugins::elements_ready_lite::ElementsReadyLiteChecker;
 use crate::checkers::wp_plugins::gtranslate::GTranslateChecker;
 use crate::checkers::wp_plugins::js_composer::JSComposerChecker;
+use crate::checkers::wp_plugins::really_simple_ssl::ReallySimpleSSLChecker;
 use crate::checkers::wp_plugins::revslider::RevSliderChecker;
 use crate::checkers::wp_plugins::woocommerce::WooCommerceChecker;
 use crate::checkers::wp_plugins::wpforms_lite::WpformsLiteChecker;
@@ -134,6 +135,7 @@ impl Application {
             Box::new(AkismetChecker::new()),
             Box::new(WpformsLiteChecker::new()),
             Box::new(AllInOneWpMigrationChecker::new()),
+            Box::new(ReallySimpleSSLChecker::new()),
         ];
 
         trace!("Returning the Application");