+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Detection of WordPress plugin Akismet.
 - Detection of WordPress plugin Wpforms lite.
 - Detection of WordPress plugin All-in-One WP Migration.
+- Detection of WordPress plugin Really Simple SSL.
 
 ### Fixed